+++ CHANGELOG.md
@@ -5,6 +5,11 @@
   ## **WORK IN PROGRESS**
   * (theimo1221) Update packages
 -->
+
+## **WORK IN PROGRESS**
+
+* (theimo1221) Allow Presence check to include acknowledgement of movement reset delay.
+
 ## 3.0.0-alpha.73 (2024-05-21)
 
 * (theimo1221) Don't send color change command on govee setLight(false) commands as color-change turns device on.